separated parts from window to widgets and constants
