+++ openpype/tools/creator/constants.py
@@ -1,0 +1,11 @@
+from Qt import QtCore
+
+
+FamilyRole = QtCore.Qt.UserRole + 3
+ExistsRole = QtCore.Qt.UserRole + 4
+PluginRole = QtCore.Qt.UserRole + 5
+
+Separator = "---separator---"
+
+# TODO regex should be defined by schema
+SubsetAllowedSymbols = "a-zA-Z0-9_."
