+++ openpype/tools/creator/widgets.py
@@ -1,4 +1,7 @@
-from Qt import QtWidgets, QtCore
+import re
+from Qt import QtWidgets, QtCore, QtGui
+
+from .constants import SubsetAllowedSymbols
 
 
 class CreateErrorMessageBox(QtWidgets.QDialog):
@@ -74,3 +77,92 @@ class CreateErrorMessageBox(QtWidgets.QDialog):
         line.setFrameShape(QtWidgets.QFrame.HLine)
         line.setFrameShadow(QtWidgets.QFrame.Sunken)
         return line
+
+
+class SubsetNameValidator(QtGui.QRegExpValidator):
+    invalid = QtCore.Signal(set)
+    pattern = "^[{}]*$".format(SubsetAllowedSymbols)
+
+    def __init__(self):
+        reg = QtCore.QRegExp(self.pattern)
+        super(SubsetNameValidator, self).__init__(reg)
+
+    def validate(self, input, pos):
+        results = super(SubsetNameValidator, self).validate(input, pos)
+        if results[0] == self.Invalid:
+            self.invalid.emit(self.invalid_chars(input))
+        return results
+
+    def invalid_chars(self, input):
+        invalid = set()
+        re_valid = re.compile(self.pattern)
+        for char in input:
+            if char == " ":
+                invalid.add("' '")
+                continue
+            if not re_valid.match(char):
+                invalid.add(char)
+        return invalid
+
+
+class SubsetNameLineEdit(QtWidgets.QLineEdit):
+    report = QtCore.Signal(str)
+    colors = {
+        "empty": (QtGui.QColor("#78879b"), ""),
+        "exists": (QtGui.QColor("#4E76BB"), "border-color: #4E76BB;"),
+        "new": (QtGui.QColor("#7AAB8F"), "border-color: #7AAB8F;"),
+    }
+
+    def __init__(self, *args, **kwargs):
+        super(SubsetNameLineEdit, self).__init__(*args, **kwargs)
+
+        validator = SubsetNameValidator()
+        self.setValidator(validator)
+        self.setToolTip("Only alphanumeric characters (A-Z a-z 0-9), "
+                        "'_' and '.' are allowed.")
+
+        self._status_color = self.colors["empty"][0]
+
+        anim = QtCore.QPropertyAnimation()
+        anim.setTargetObject(self)
+        anim.setPropertyName(b"status_color")
+        anim.setEasingCurve(QtCore.QEasingCurve.InCubic)
+        anim.setDuration(300)
+        anim.setStartValue(QtGui.QColor("#C84747"))  # `Invalid` status color
+        self.animation = anim
+
+        validator.invalid.connect(self.on_invalid)
+
+    def on_invalid(self, invalid):
+        message = "Invalid character: %s" % ", ".join(invalid)
+        self.report.emit(message)
+        self.animation.stop()
+        self.animation.start()
+
+    def as_empty(self):
+        self._set_border("empty")
+        self.report.emit("Empty subset name ..")
+
+    def as_exists(self):
+        self._set_border("exists")
+        self.report.emit("Existing subset, appending next version.")
+
+    def as_new(self):
+        self._set_border("new")
+        self.report.emit("New subset, creating first version.")
+
+    def _set_border(self, status):
+        qcolor, style = self.colors[status]
+        self.animation.setEndValue(qcolor)
+        self.setStyleSheet(style)
+
+    def _get_status_color(self):
+        return self._status_color
+
+    def _set_status_color(self, color):
+        self._status_color = color
+        self.setStyleSheet("border-color: %s;" % color.name())
+
+    status_color = QtCore.Property(
+        QtGui.QColor, _get_status_color, _set_status_color
+    )
